Move analysis_module_flag_enum to ies_config
