+++ libres/lib/include/ert/analysis/analysis_module.hpp
@@ -19,18 +19,9 @@
 #ifndef ERT_ANALYSIS_MODULE_H
 #define ERT_ANALYSIS_MODULE_H
 
-#include <ert/util/bool_vector.hpp>
-
 #include <ert/analysis/ies/ies_data.hpp>
 #include <ert/analysis/ies/ies_config.hpp>
 
-enum analysis_module_flag_enum : int {
-    ANALYSIS_USE_A =
-        4, // The module will read the content of A - but not modify it.
-    ANALYSIS_UPDATE_A =
-        8, // The update will be based on modifying A directly, and not on an X matrix.
-    ANALYSIS_ITERABLE = 32 // The module can bu used as an iterative smoother.
-};
 
 typedef enum {
     ENSEMBLE_SMOOTHER = 1,
